Refactor map view so can be for vets or salons
